refactor(tests): update app/cardList/header/brokenComponent
